requirements.txt
# core libraries
torch>=2.2.0
torchvision>=0.16.0
tensorflow==2.15.0

# supporting libraries
transformers==4.40.1
tokenizers==0.19.1
sqlmodel
timm==0.9.10
Pillow
imageio
litellm
opencv-python
python-Levenshtein
sentencepiece
einops
pycocotools
sentence-transformers
tenacity
modal

# datasets and utilities
datasets
tensorflow-datasets==4.9.3
moviepy==1.0.3
pymongo

# llm providers
vertexai

# data viz
streamlit
pandas
plotly
umap-learn
hdbscan
kaleido
pdfkit
matplotlib

# database
faiss-cpu #TODO: try faiss-gpu